basic-layout/src/main/java/org/websandbox/learning/layout/user/UserServiceImpl.java
package org.websandbox.learning.layout.user;

import java.util.List;
import java.util.Optional;

import javax.transaction.Transactional;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
@Transactional
public class UserServiceImpl implements UserService {

	@Autowired
	UserRepository userRepository;

	@Override
	public Optional<User> getUserById(Long id) {
		return userRepository.findById(id);
	}

	@Override
	public Optional<User> getUserByName(String name) {
		return userRepository.findByName(name);
	}

	@Override
	public boolean exists(String name) {
		if (userRepository.findByName(name) != null) {
			return true;
		}
		return false;
	}

	@Override
	public User save(User user) {
		return userRepository.save(user);
	}

	@Override
	public List<User> getAllUsers() {
		return userRepository.findAll();
	}
}
